requirements.txt
customtkinter>=5.2.0
python-docx>=0.8.11
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-multipart>=0.0.9
